Fix incorrect `hasAssertions` duplication
Vitest has a method `expect.assertions` which can be used to expect a
specific amount of assertions and it has a method `expect.hasAssertions`
to wait for any number of assertions (>=1).

We rename the methods to match the methods in Vitest itself so that the
documentation works. The only difference is that in the ReScript
bindings we have `suite` to double as the object `expect` since a
function and object can't share a name in ReScript.