Turn bindings into zero-cost
This makes a few API breaking changes and removes some of the Functor
magic in favor of duplication. The result is that the test bindings
themselves become slightly more verbose in tests but are also
zero-cost.

This fixes an issue where Vitest would complain that it was imported
from a dependency of the test which causes the set-up to be performed
multiple times:

```
Error: Vitest was initialized with native Node instead of Vite Node.

One of the following is possible:
- "vitest" is imported outside of your tests (in that case, use "vitest/node" or import.meta.vitest)
- "vitest" is imported inside "globalSetup" (use "setupFiles", because "globalSetup" runs in a different context)
- Your dependency inside "node_modules" imports "vitest" directly (in that case, inline that dependency, using "deps.inline" config)
- Otherwise, it might be a Vitest bug. Please report it to https://github.com/vitest-dev/vitest/issues
```